==> Scripts/03_GrowthCurve_modelo.R <==
################################################################################
################ SbIII promastigote Dose-response analysis #####################
################################################################################

library(dplyr)
library(ggplot2) 
library(drc)
library(gridExtra)
# Loading data 

DataGC <- read.csv(file = "Data/Processed/DataGC.csv")

# Checking data
head(DataGC)
sapply(DataGC, class)
DataGC$experiment <- as.factor(DataGC$experiment)
DataGC$pop <- as.factor(DataGC$pop)
sapply(DataGC, class)

# Removing time 0
DataGC <-  DataGC[DataGC$tempo != 1, ]
DataGC <- DataGC%>%
group_by(conc, pop, experiment, tempo)%>%
summarise(mean_value = mean(abs), sd_value = sd(abs))

# # DataGC_SUM <- rename(DataGC, conc = "conc", pop = "pop",
#                      experiment = "experiment")
# 
# # Adding mean values


# DataGC <- left_join(DataGC, sum_values, by = c("conc", "pop"))


#Splitting populations per conc

REF_C0 <- filter(DataGC, pop == "REF", conc =="0")
GSH1_C0 <- filter(DataGC, pop == "GSH1",conc =="0")
C6_C0 <- filter(DataGC, pop == "C6",conc =="0")
C7_C0 <- filter(DataGC, pop == "C7",conc =="0")
C44_C0 <- filter(DataGC, pop == "C44",conc =="0")
C58_C0 <- filter(DataGC, pop == "C58",conc =="0")
C67_C0 <- filter(DataGC, pop == "C67",conc =="0")
C73_C0 <- filter(DataGC, pop == "C73",conc =="0")
C85_C0 <- filter(DataGC, pop == "C85",conc =="0")
C89_C0 <- filter(DataGC, pop == "C89",conc =="0")

# Adjusting the models

REF_C0_G4 <- drm(mean_value ~ tempo, data = REF_C0 , fct =  G.4())
plot(REF_C0_G4, 
     broken = TRUE, gridsize = 100, xlab = "Tempo(dias)", 
     ylab = "Abs", type = "all")
summary(REF_C0_G4)

GSH1_C0_G4 <- drm(mean_value ~ tempo, data = GSH1_C0 , fct =  G.4())
plot(GSH1_C0_G4, broken = TRUE, gridsize = 100, xlab = "Tempo (dias)", 
     ylab = "Abs", type = "all")

C6_C0_G4 <- drm(mean_value ~ tempo, data = C6_C0 , fct =  G.4())
plot(C6_C0_G4, broken = TRUE, gridsize = 100, xlab = "Tempo (dias)", 
     ylab = "Abs", type = "all")

C7_C0_G4 <- drm(mean_value ~ tempo, data = C7_C0 , fct =  G.4())
plot(C7_C0_G4, broken = TRUE, gridsize = 100, xlab = "Tempo (dias)", 
     ylab = "Abs", type = "all")

C44_C0_G4 <- drm(mean_value ~ tempo, data = C44_C0 , fct =  G.4())
plot(C44_C0_G4, broken = TRUE, gridsize = 100, xlab = "Tempo (dias)", 
     ylab = "Abs", type = "all")

C58_C0_G4 <- drm(mean_value ~ tempo, data = C58_C0 , fct =  G.4())
plot(C58_C0_G4, broken = TRUE, gridsize = 100, xlab = "Tempo (dias)", 
     ylab = "Abs", type = "all")


C67_C0_G4 <- drm(mean_value ~ tempo, data = C67_C0 , fct =  G.4())
plot(C67_C0_G4, broken = TRUE, gridsize = 100, xlab = "Tempo (dias)", 
     ylab = "Abs", type = "all")

C73_C0_G4 <- drm(mean_value ~ tempo, data = C73_C0 , fct =  G.4())
plot(C73_C0_G4, broken = TRUE, gridsize = 100, xlab = "Tempo (dias)", 
     ylab = "Abs", type = "all")

C85_C0_G4 <- drm(mean_value ~ tempo, data = C85_C0 , fct =  G.4())
plot(C85_C0_G4, broken = TRUE, gridsize = 100, xlab = "Tempo (dias)", 
     ylab = "Abs", type = "all")

C89_C0_G4 <- drm(mean_value ~ tempo, data = C89_C0 , fct =  G.4())
plot(C89_C0_G4, broken = TRUE, gridsize = 100, xlab = "Tempo (dias)", 
     ylab = "Abs", type = "all")


# C0 analysis

newdata_REF_C0 <- expand.grid(tempo = seq(min(REF_C0$tempo), max(REF_C0$tempo), length.out = 200))
newdata_REF_C0$abs <- predict(REF_C0_G4, newdata_REF_C0, type = 'response')
newdata_REF_C0$pop <- "REF"

newdata_GSH1_C0 <- expand.grid(tempo = seq(min(GSH1_C0$tempo), max(GSH1_C0$tempo), length.out = 200))
newdata_GSH1_C0$abs <- predict(GSH1_C0_G4, newdata_GSH1_C0, type = 'response')
newdata_GSH1_C0$pop <- "GSH1"

newdata_C6_C0 <- expand.grid(tempo = seq(min(C6_C0$tempo), max(C6_C0$tempo), length.out = 200))
newdata_C6_C0$abs <- predict(C6_C0_G4, newdata_C6_C0, type = 'response')
newdata_C6_C0$pop <- "C6"

newdata_C7_C0 <- expand.grid(tempo = seq(min(C7_C0$tempo), max(C7_C0$tempo), length.out = 200))
newdata_C7_C0$abs <- predict(C7_C0_G4, newdata_C7_C0, type = 'response')
newdata_C7_C0$pop <- "C7"

newdata_C44_C0 <- expand.grid(tempo = seq(min(C44_C0$tempo), max(C44_C0$tempo), length.out = 200))
newdata_C44_C0$abs <- predict(C44_C0_G4, newdata_C44_C0, type = 'response')
newdata_C44_C0$pop <- "C44"

newdata_C58_C0 <- expand.grid(tempo = seq(min(C58_C0$tempo), max(C58_C0$tempo), length.out = 200))
newdata_C58_C0$abs <- predict(C58_C0_G4, newdata_C58_C0, type = 'response')
newdata_C58_C0$pop <- "C58"

newdata_C67_C0 <- expand.grid(tempo = seq(min(C67_C0$tempo), max(C67_C0$tempo), length.out = 200))
newdata_C67_C0$abs <- predict(C67_C0_G4, newdata_C67_C0, type = 'response')
newdata_C67_C0$pop <- "C67"

newdata_C73_C0 <- expand.grid(tempo = seq(min(C73_C0$tempo), max(C73_C0$tempo), length.out = 200))
newdata_C73_C0$abs <- predict(C73_C0_G4, newdata_C73_C0, type = 'response')
newdata_C73_C0$pop <- "C73"

newdata_C85_C0 <- expand.grid(tempo = seq(min(C85_C0$tempo), max(C85_C0$tempo), length.out = 200))
newdata_C85_C0$abs <- predict(C85_C0_G4, newdata_C85_C0, type = 'response')
newdata_C85_C0$pop <- "C85"

newdata_C89_C0 <- expand.grid(tempo = seq(min(C89_C0$tempo), max(C89_C0$tempo), length.out = 200))
newdata_C89_C0$abs <- predict(C89_C0_G4, newdata_C89_C0, type = 'response')
newdata_C89_C0$pop <- "C89"

## Exploring predicted model

GrowthCurves_C0 <- full_join(REF_C0, C6_C0)
GrowthCurves_C0 <- full_join(GrowthCurves_C0, GSH1_C0)
GrowthCurves_C0 <- full_join(GrowthCurves_C0, C7_C0)
GrowthCurves_C0 <- full_join(GrowthCurves_C0, C44_C0)
GrowthCurves_C0 <- full_join(GrowthCurves_C0, C58_C0)
GrowthCurves_C0 <- full_join(GrowthCurves_C0, C67_C0)
GrowthCurves_C0 <- full_join(GrowthCurves_C0, C73_C0)
GrowthCurves_C0 <- full_join(GrowthCurves_C0, C85_C0)
GrowthCurves_C0 <- full_join(GrowthCurves_C0, C89_C0)

newdata_full <- rbind.data.frame(newdata_REF_C0, newdata_C6_C0)
newdata_full <- rbind(newdata_full, newdata_GSH1_C0)
newdata_full <- rbind(newdata_full, newdata_C7_C0)
newdata_full <- rbind(newdata_full, newdata_C44_C0)
newdata_full <- rbind(newdata_full, newdata_C58_C0)
newdata_full <- rbind(newdata_full, newdata_C67_C0)
newdata_full <- rbind(newdata_full, newdata_C73_C0)
newdata_full <- rbind(newdata_full, newdata_C85_C0)
newdata_full <- rbind(newdata_full, newdata_C89_C0)

DoseResponseCurves_01 <- ggplot() +
  geom_point(data = REF_C0, aes(x = tempo, y = mean_value)) +
  geom_line(data = newdata_REF_C0, aes(x = tempo, y = newdata_REF_C0$abs)) +
  geom_point(data = GSH1_C0, aes(x = tempo, y = mean_value)) +
  geom_line(data = newdata_GSH1_C0, aes(x = tempo, y = newdata_GSH1_C0$abs))+
  geom_point(data = C6_C0, aes(x = tempo, y = mean_value)) +
  geom_line(data = newdata_C6_C0, aes(x = tempo, y = newdata_C6_C0$abs))+
  geom_point(data = C7_C0, aes(x = tempo, y = mean_value)) +
  geom_line(data = newdata_C7_C0, aes(x = tempo, y = newdata_C7_C0$abs)) +
  geom_point(data = C44_C0, aes(x = tempo, y = mean_value)) +
  geom_line(data = newdata_C44_C0
            , aes(x = tempo, y = newdata_C44_C0$abs))+
  geom_point(data = C58_C0, aes(x = tempo, y = mean_value)) +
  geom_line(data = newdata_C58_C0, aes(x = tempo, y = newdata_C58_C0$abs))+
  geom_point(data = C67_C0, aes(x = tempo, y = mean_value)) +
  geom_line(data = newdata_C67_C0, aes(x = tempo, y =  newdata_C67_C0$abs )) +
  geom_point(data = C73_C0, aes(x = tempo, y = mean_value)) +
  geom_line(data = newdata_C73_C0, aes(x = tempo, y =newdata_C73_C0$abs )) +
  geom_point(data = C85_C0, aes(x = tempo, y = mean_value)) +
  geom_line(data = newdata_C85_C0, aes(x = tempo, y = newdata_C85_C0$abs))+
  geom_point(data = C89_C0, aes(x = tempo, y = mean_value)) +
  geom_line(data = newdata_C89_C0, aes(x = tempo, y = newdata_C89_C0$abs))+
  ggtitle("Curva ajustada por clone") +
  labs(x = "Tempo (dias)", y = "Abs (600nm)") +
  theme_bw() +
  theme( plot.title = element_text(size = 14, face = "bold"),
         axis.text.x = element_text(size = 10),
         axis.text.y = element_text(size = 10),
         axis.title.x = element_text(size = 15),
         axis.title.y = element_text(size = 15)) +
  facet_wrap(~pop, nrow = 2)

DoseResponseCurves_01
ggsave("Figuras/01_DoseResponseCurves_CONC_0.png")

summary(REF_C0_G4)
# Extracting measures

COEFICIENTS <- c("GrowthRate", "c", "d", "e")
modelREF <- summary(REF_C0_G4)
confintREF <- confint(REF_C0_G4)
modelREF_coef <- as.data.frame(modelREF$coefficients)
modelREF_conf <- as.data.frame(confintREF)
modelREF_SUM <- cbind(modelREF_coef, modelREF_conf)
modelREF_SUM$coef <- COEFICIENTS 
modelREF_SUM$pop <- rep("REF_C0", 4)

COEFICIENTS <- c("GrowthRate", "c", "d", "e")
modelGSH1 <- summary(GSH1_C0_G4)
confintGSH1 <- confint(GSH1_C0_G4)
modelGSH1_coef <- as.data.frame(modelGSH1$coefficients)
modelGSH1_conf <- as.data.frame(confintGSH1)
modelGSH1_SUM <- cbind(modelGSH1_coef, modelGSH1_conf)
modelGSH1_SUM$coef <- COEFICIENTS 
modelGSH1_SUM$pop <- rep("GSH1_C0", 4)

COEFICIENTS <- c("GrowthRate", "c", "d", "e")
modelC6 <- summary(C6_C0_G4)
confintC6 <- confint(C6_C0_G4)
modelC6_coef <- as.data.frame(modelC6$coefficients)
modelC6_conf <- as.data.frame(confintC6)
modelC6_SUM <- cbind(modelC6_coef, modelC6_conf)
modelC6_SUM$coef <- COEFICIENTS 
modelC6_SUM$pop <- rep("C6_C0", 4)

COEFICIENTS <- c("GrowthRate", "c", "d", "e")
modelC7 <- summary(C7_C0_G4)
confintC7 <- confint(C7_C0_G4)
modelC7_coef <- as.data.frame(modelC7$coefficients)
modelC7_conf <- as.data.frame(confintC7)
modelC7_SUM <- cbind(modelC7_coef, modelC7_conf)
modelC7_SUM$coef <- COEFICIENTS 
modelC7_SUM$pop <- rep("C7_C0", 4)

COEFICIENTS <- c("GrowthRate", "c", "d", "e")
modelC44 <- summary(C44_C0_G4)
confintC44 <- confint(C44_C0_G4)
modelC44_coef <- as.data.frame(modelC44$coefficients)
modelC44_conf <- as.data.frame(confintC44)
modelC44_SUM <- cbind(modelC44_coef, modelC44_conf)
modelC44_SUM$coef <- COEFICIENTS 
modelC44_SUM$pop <- rep("C44_C0", 4)

COEFICIENTS <- c("GrowthRate", "c", "d", "e")
modelC58 <- summary(C58_C0_G4)
confintC58 <- confint(C58_C0_G4)
modelC58_coef <- as.data.frame(modelC58$coefficients)
modelC58_conf <- as.data.frame(confintC58)
modelC58_SUM <- cbind(modelC58_coef, modelC58_conf)
modelC58_SUM$coef <- COEFICIENTS 
modelC58_SUM$pop <- rep("C58_C0", 4)

COEFICIENTS <- c("GrowthRate", "c", "d", "e")
modelC67 <- summary(C67_C0_G4)
confintC67 <- confint(C67_C0_G4)
modelC67_coef <- as.data.frame(modelC67$coefficients)
modelC67_conf <- as.data.frame(confintC67)
modelC67_SUM <- cbind(modelC67_coef, modelC67_conf)
modelC67_SUM$coef <- COEFICIENTS 
modelC67_SUM$pop <- rep("C67_C0", 4)

COEFICIENTS <- c("GrowthRate", "c", "d", "e")
modelC73 <- summary(C73_C0_G4)
confintC73 <- confint(C73_C0_G4)
modelC73_coef <- as.data.frame(modelC73$coefficients)
modelC73_conf <- as.data.frame(confintC73)
modelC73_SUM <- cbind(modelC73_coef, modelC73_conf)
modelC73_SUM$coef <- COEFICIENTS 
modelC73_SUM$pop <- rep("C73_C0", 4)

COEFICIENTS <- c("GrowthRate", "c", "d", "e")
modelC85 <- summary(C85_C0_G4)
confintC85 <- confint(C85_C0_G4)
modelC85_coef <- as.data.frame(modelC85$coefficients)
modelC85_conf <- as.data.frame(confintC85)
modelC85_SUM <- cbind(modelC85_coef, modelC85_conf)
modelC85_SUM$coef <- COEFICIENTS 
modelC85_SUM$pop <- rep("C85_C0", 4)

COEFICIENTS <- c("GrowthRate", "c", "d", "e")
modelC89 <- summary(C89_C0_G4)
confintC89 <- confint(C89_C0_G4)
modelC89_coef <- as.data.frame(modelC89$coefficients)
modelC89_conf <- as.data.frame(confintC89)
modelC89_SUM <- cbind(modelC89_coef, modelC89_conf)
modelC89_SUM$coef <- COEFICIENTS 
modelC89_SUM$pop <- rep("C89_C0", 4)


summary_DR <- rbind(modelREF_SUM,modelGSH1_SUM,
                    modelC7_SUM, modelC6_SUM, modelC44_SUM,
                    modelC58_SUM, modelC67_SUM, modelC73_SUM,
                    modelC85_SUM, modelC89_SUM)

write.csv(summary_DR , file = "Docs/summary_GC_GSH1_C0.csv", row.names = FALSE)

# Models comparisons

m1<-drm(mean_value ~ tempo, pop, data = DataGC, 
        fct =G.4(names = c("GrowthRate", "c", "d", "e")))
m2<-drm(mean_value ~ tempo, data = DataGC, 
        fct =G.4(names = c("GrowthRate", "c", "d", "e")))

anova(m1,m2) 

DoseResponseCurves_02 <- ggplot() +
  geom_point(data = REF_C0, aes(x = tempo, y = mean_value, color = "REF")) +
  geom_line(data = newdata_REF_C0, aes(x = tempo, y = newdata_REF_C0$abs, color = "REF")) +
  geom_point(data = GSH1_C0, aes(x = tempo, y = mean_value, color = "GSH1")) +
  geom_line(data = newdata_GSH1_C0, aes(x = tempo, y = newdata_GSH1_C0$abs, color = "GSH1")) +
  geom_point(data = C6_C0, aes(x = tempo, y = mean_value, color = "C6")) +
  geom_line(data = newdata_C6_C0, aes(x = tempo, y = newdata_C6_C0$abs, color = "C6")) +
  geom_point(data = C7_C0, aes(x = tempo, y = mean_value, color = "C7")) +
  geom_line(data = newdata_C7_C0, aes(x = tempo, y = newdata_C7_C0$abs, color = "C7")) +
  geom_point(data = C44_C0, aes(x = tempo, y = mean_value, color = "C44")) +
  geom_line(data = newdata_C44_C0, aes(x = tempo, y = newdata_C44_C0$abs, color = "C44")) +
  geom_point(data = C58_C0, aes(x = tempo, y = mean_value, color = "C58")) +
  geom_line(data = newdata_C58_C0, aes(x = tempo, y = newdata_C58_C0$abs, color = "C58")) +
  geom_point(data = C67_C0, aes(x = tempo, y = mean_value, color = "C67")) +
  geom_line(data = newdata_C67_C0, aes(x = tempo, y = newdata_C67_C0$abs, color = "C67")) +
  geom_point(data = C73_C0, aes(x = tempo, y = mean_value, color = "C73")) +
  geom_line(data = newdata_C73_C0, aes(x = tempo, y = newdata_C73_C0$abs, color = "C73")) +
  geom_point(data = C85_C0, aes(x = tempo, y = mean_value, color = "C85")) +
  geom_line(data = newdata_C85_C0, aes(x = tempo, y = newdata_C85_C0$abs, color = "C85")) +
  geom_point(data = C89_C0, aes(x = tempo, y = mean_value, color = "C89")) +
  geom_line(data = newdata_C89_C0, aes(x = tempo, y = newdata_C89_C0$abs, color = "C89")) +
  ggtitle("Promastigotes Growth") +
  labs(x = "Tempo (dias)", y = "Abs", color = "População") +
  theme_bw() +
  theme(
    plot.title = element_text(size = 14, face = "bold"),
    axis.text.x = element_text(size = 14, face = "bold"),
    axis.text.y = element_text(size = 14, face = "bold"),
    axis.title.x = element_text(size = 14, face = "bold"),
    axis.title.y = element_text(size = 14, face = "bold")
  )

DoseResponseCurves_02

ggsave("Figuras/08_DoseResponseCurves_02.png")

DoseResponseCurves_02+ labs(color = "Populações", size = 25)+
  theme(plot.title = element_text(size = 20, face = "bold"),
        axis.text.x = element_text(size = 20, face = "bold"),
        axis.text.y = element_text(size = 20, face = "bold"),
        axis.title.x = element_text(size = 20, face = "bold"),
        axis.title.y = element_text(size = 20, face = "bold"),
        legend.text = element_text(size = 20),
        legend.title = element_text(size = 25,  face = "bold"),
        legend.position = c(0.95,0.8),   
        legend.justification = "right",
        legend.box.background = element_blank())

ggsave("Figures/02_DoseResponseCurves_CONC_0.png")


#  Fit diagnostic
plot1 <- ggplot(REF_C0, aes(x = fitted(REF_C0_G4), y = mean_value)) +
  geom_point(pch = 16, cex = 0.7) +
  geom_abline(intercept = 0, slope = 1) +
  labs(title = " REF - G4 - Ajustados x Observados", x = "Valores ajustados", 
       y = "Valores observados") +
  theme_minimal()

plot2 <- ggplot(REF_C0, aes(x = fitted(REF_C0_G4), y = residuals(REF_C0_G4, type = "studentised"))) +
  geom_point(pch = 16, cex = 0.7) +
  geom_smooth(method = "loess", se = FALSE, color = "blue") +
  geom_hline(yintercept = 0, linetype = "dashed") +
  labs(title = "REF - G4 - Resíduos x Ajustados", x = "Valores ajustados", 
       y = "Resíduos") +
  theme_minimal()

plot3 <- ggplot(REF_C0, aes(sample = residuals(REF_C0_G4, type = "studentised"))) +
  geom_qq() +
  geom_qq_line() +
  labs(title = "G4", x = "Quantis teóricos", y = "Resíduos") +
  theme_minimal()

dfit_plot_1 <- grid.arrange(plot1, plot2, plot3, ncol = 3)

ggsave("Figuras/Fit_REF.png", plot)

plot1 <- ggplot(GSH1_C0, aes(x = fitted(GSH1_C0_G4), y = mean_value)) +
  geom_point(pch = 16, cex = 0.7) +
  geom_abline(intercept = 0, slope = 1) +
  labs(title = " GSH1 - G4 - Ajustados x Observados", x = "Valores ajustados", 
       y = "Valores observados") +
  theme_minimal()

plot2 <- ggplot(GSH1_C0, aes(x = fitted(GSH1_C0_G4), y = residuals(GSH1_C0_G4, type = "studentised"))) +
  geom_point(pch = 16, cex = 0.7) +
  geom_smooth(method = "loess", se = FALSE, color = "blue") +
  geom_hline(yintercept = 0, linetype = "dashed") +
  labs(title = "GSH1 - G4 - Resíduos x Ajustados", x = "Valores ajustados", 
       y = "Resíduos") +
  theme_minimal()

plot3 <- ggplot(GSH1_C0, aes(sample = residuals(GSH1_C0_G4, type = "studentised"))) +
  geom_qq() +
  geom_qq_line() +
  labs(title = "G4", x = "Quantis teóricos", y = "Resíduos") +
  theme_minimal()

dfit_plot_2 <- grid.arrange(plot1, plot2, plot3, ncol = 3)

dfit_plot_2
ggsave("Figuras/Fit_GSH1.png", plot)

plot1 <- ggplot(C6_C0, aes(x = fitted(C6_C0_G4), y = mean_value)) +
  geom_point(pch = 16, cex = 0.7) +
  geom_abline(intercept = 0, slope = 1) +
  labs(title = " C6 - G4 - Ajustados x Observados", x = "Valores ajustados", 
       y = "Valores observados") +
  theme_minimal()

plot2 <- ggplot(C6_C0, aes(x = fitted(C6_C0_G4), y = residuals(C6_C0_G4, type = "studentised"))) +
  geom_point(pch = 16, cex = 0.7) +
  geom_smooth(method = "loess", se = FALSE, color = "blue") +
  geom_hline(yintercept = 0, linetype = "dashed") +
  labs(title = "C6 - G4 - Resíduos x Ajustados", x = "Valores ajustados", 
       y = "Resíduos") +
  theme_minimal()

plot3 <- ggplot(C6_C0, aes(sample = residuals(C6_C0_G4, type = "studentised"))) +
  geom_qq() +
  geom_qq_line() +
  labs(title = "G4", x = "Quantis teóricos", y = "Resíduos") +
  theme_minimal()

dfit_plot <- grid.arrange(plot1, plot2, plot3, ncol = 3)

ggsave("Figuras/Fit_C6.png", plot)

plot1 <- ggplot(C7_C0, aes(x = fitted(C7_C0_G4), y = mean_value)) +
  geom_point(pch = 16, cex = 0.7) +
  geom_abline(intercept = 0, slope = 1) +
  labs(title = " C7 - G4 - Ajustados x Observados", x = "Valores ajustados", 
       y = "Valores observados") +
  theme_minimal()

plot2 <- ggplot(C7_C0, aes(x = fitted(C7_C0_G4), y = residuals(C7_C0_G4, type = "studentised"))) +
  geom_point(pch = 16, cex = 0.7) +
  geom_smooth(method = "loess", se = FALSE, color = "blue") +
  geom_hline(yintercept = 0, linetype = "dashed") +
  labs(title = "C7 - G4 - Resíduos x Ajustados", x = "Valores ajustados", 
       y = "Resíduos") +
  theme_minimal()

plot3 <- ggplot(C7_C0, aes(sample = residuals(C7_C0_G4, type = "studentised"))) +
  geom_qq() +
  geom_qq_line() +
  labs(title = "G4", x = "Quantis teóricos", y = "Resíduos") +
  theme_minimal()

dfit_plot <- grid.arrange(plot1, plot2, plot3, ncol = 3)

ggsave("Figuras/Fit_C7.png", plot)

plot1 <- ggplot(C44_C0, aes(x = fitted(C44_C0_G4), y = mean_value)) +
  geom_point(pch = 16, cex = 0.7) +
  geom_abline(intercept = 0, slope = 1) +
  labs(title = " C44 - G4 - Ajustados x Observados", x = "Valores ajustados", 
       y = "Valores observados") +
  theme_minimal()

plot2 <- ggplot(C44_C0, aes(x = fitted(C44_C0_G4), y = residuals(C44_C0_G4, type = "studentised"))) +
  geom_point(pch = 16, cex = 0.7) +
  geom_smooth(method = "loess", se = FALSE, color = "blue") +
  geom_hline(yintercept = 0, linetype = "dashed") +
  labs(title = "C44 - G4 - Resíduos x Ajustados", x = "Valores ajustados", 
       y = "Resíduos") +
  theme_minimal()

plot3 <- ggplot(C44_C0, aes(sample = residuals(C44_C0_G4, type = "studentised"))) +
  geom_qq() +
  geom_qq_line() +
  labs(title = "G4", x = "Quantis teóricos", y = "Resíduos") +
  theme_minimal()

dfit_plot <- grid.arrange(plot1, plot2, plot3, ncol = 3)

ggsave("Figuras/Fit_C44.png", plot)

plot1 <- ggplot(C58_C0, aes(x = fitted(C58_C0_G4), y = mean_value)) +
  geom_point(pch = 16, cex = 0.7) +
  geom_abline(intercept = 0, slope = 1) +
  labs(title = " C58 - G4 - Ajustados x Observados", x = "Valores ajustados", 
       y = "Valores observados") +
  theme_minimal()

plot2 <- ggplot(C58_C0, aes(x = fitted(C58_C0_G4), y = residuals(C58_C0_G4, type = "studentised"))) +
  geom_point(pch = 16, cex = 0.7) +
  geom_smooth(method = "loess", se = FALSE, color = "blue") +
  geom_hline(yintercept = 0, linetype = "dashed") +
  labs(title = "C58 - G4 - Resíduos x Ajustados", x = "Valores ajustados", 
       y = "Resíduos") +
  theme_minimal()

plot3 <- ggplot(C58_C0, aes(sample = residuals(C58_C0_G4, type = "studentised"))) +
  geom_qq() +
  geom_qq_line() +
  labs(title = "G4", x = "Quantis teóricos", y = "Resíduos") +
  theme_minimal()

dfit_plot <- grid.arrange(plot1, plot2, plot3, ncol = 3)

ggsave("Figuras/Fit_C58.png", plot)

plot1 <- ggplot(C67_C0, aes(x = fitted(C67_C0_G4), y = mean_value)) +
  geom_point(pch = 16, cex = 0.7) +
  geom_abline(intercept = 0, slope = 1) +
  labs(title = " C67 - G4 - Ajustados x Observados", x = "Valores ajustados", 
       y = "Valores observados") +
  theme_minimal()

plot2 <- ggplot(C67_C0, aes(x = fitted(C67_C0_G4), y = residuals(C67_C0_G4, type = "studentised"))) +
  geom_point(pch = 16, cex = 0.7) +
  geom_smooth(method = "loess", se = FALSE, color = "blue") +
  geom_hline(yintercept = 0, linetype = "dashed") +
  labs(title = "C67 - G4 - Resíduos x Ajustados", x = "Valores ajustados", 
       y = "Resíduos") +
  theme_minimal()

plot3 <- ggplot(C67_C0, aes(sample = residuals(C67_C0_G4, type = "studentised"))) +
  geom_qq() +
  geom_qq_line() +
  labs(title = "G4", x = "Quantis teóricos", y = "Resíduos") +
  theme_minimal()

dfit_plot <- grid.arrange(plot1, plot2, plot3, ncol = 3)

ggsave("Figures/Fit_C67.png", plot)


plot1 <- ggplot(C73_C0, aes(x = fitted(C73_C0_G4), y = mean_value)) +
  geom_point(pch = 16, cex = 0.7) +
  geom_abline(intercept = 0, slope = 1) +
  labs(title = " C73 - G4 - Ajustados x Observados", x = "Valores ajustados", 
       y = "Valores observados") +
  theme_minimal()

plot2 <- ggplot(C73_C0, aes(x = fitted(C73_C0_G4), y = residuals(C73_C0_G4, type = "studentised"))) +
  geom_point(pch = 16, cex = 0.7) +
  geom_smooth(method = "loess", se = FALSE, color = "blue") +
  geom_hline(yintercept = 0, linetype = "dashed") +
  labs(title = "C73 - G4 - Resíduos x Ajustados", x = "Valores ajustados", 
       y = "Resíduos") +
  theme_minimal()

plot3 <- ggplot(C73_C0, aes(sample = residuals(C73_C0_G4, type = "studentised"))) +
  geom_qq() +
  geom_qq_line() +
  labs(title = "G4", x = "Quantis teóricos", y = "Resíduos") +
  theme_minimal()

dfit_plot <- grid.arrange(plot1, plot2, plot3, ncol = 3)

ggsave("Figures/Fit_C73.png", plot)

plot1 <- ggplot(C85_C0, aes(x = fitted(C85_C0_G4), y = mean_value)) +
  geom_point(pch = 16, cex = 0.7) +
  geom_abline(intercept = 0, slope = 1) +
  labs(title = " C85 - G4 - Ajustados x Observados", x = "Valores ajustados", 
       y = "Valores observados") +
  theme_minimal()

plot2 <- ggplot(C85_C0, aes(x = fitted(C85_C0_G4), y = residuals(C85_C0_G4, type = "studentised"))) +
  geom_point(pch = 16, cex = 0.7) +
  geom_smooth(method = "loess", se = FALSE, color = "blue") +
  geom_hline(yintercept = 0, linetype = "dashed") +
  labs(title = "C85 - G4 - Resíduos x Ajustados", x = "Valores ajustados", 
       y = "Resíduos") +
  theme_minimal()

plot3 <- ggplot(C85_C0, aes(sample = residuals(C85_C0_G4, type = "studentised"))) +
  geom_qq() +
  geom_qq_line() +
  labs(title = "G4", x = "Quantis teóricos", y = "Resíduos") +
  theme_minimal()

dfit_plot <- grid.arrange(plot1, plot2, plot3, ncol = 3)

ggsave("Figures/Fit_C85.png", plot)


plot1 <- ggplot(C89_C0, aes(x = fitted(C89_C0_G4), y = mean_value)) +
  geom_point(pch = 16, cex = 0.7) +
  geom_abline(intercept = 0, slope = 1) +
  labs(title = " C89 - G4 - Ajustados x Observados", x = "Valores ajustados", 
       y = "Valores observados") +
  theme_minimal()

plot2 <- ggplot(C89_C0, aes(x = fitted(C89_C0_G4), y = residuals(C89_C0_G4, type = "studentised"))) +
  geom_point(pch = 16, cex = 0.7) +
  geom_smooth(method = "loess", se = FALSE, color = "blue") +
  geom_hline(yintercept = 0, linetype = "dashed") +
  labs(title = "C89 - G4 - Resíduos x Ajustados", x = "Valores ajustados", 
       y = "Resíduos") +
  theme_minimal()

plot3 <- ggplot(C89_C0, aes(sample = residuals(C89_C0_G4, type = "studentised"))) +
  geom_qq() +
  geom_qq_line() +
  labs(title = "G4", x = "Quantis teóricos", y = "Resíduos") +
  theme_minimal()

dfit_plot <- grid.arrange(plot1, plot2, plot3, ncol = 3)

ggsave("Figures/Fit_C89.png", plot)

# C150 ANALYSIS

# Splitting populations per conc

REF_C150 <- filter(DataGC, pop == "REF", conc =="150")
GSH1_C150 <- filter(DataGC, pop == "GSH1",conc =="150")
C6_C150 <- filter(DataGC, pop == "C6",conc =="150")
C7_C150 <- filter(DataGC, pop == "C7",conc =="150")
C44_C150 <- filter(DataGC, pop == "C44",conc =="150")
C58_C150 <- filter(DataGC, pop == "C58",conc =="150")
C67_C150 <- filter(DataGC, pop == "C67",conc =="150")
C73_C150 <- filter(DataGC, pop == "C73",conc =="150")
C85_C150 <- filter(DataGC, pop == "C85",conc =="150")
C89_C150 <- filter(DataGC, pop == "C89",conc =="150")

# Models 150

REF_C150_G4 <- drm(mean_value ~ tempo, data = REF_C150 , fct =  G.4())
plot(REF_C150_G4, 
     broken = TRUE, gridsize = 100, xlab = "Tempo(dias)", 
     ylab = "Abs", type = "all")
summary(REF_C150_G4)
GSH1_C150_G4 <- drm(mean_value~ tempo, data = GSH1_C150 , fct =  G.4())
plot(GSH1_C150_G4, broken = TRUE, gridsize = 100, xlab = "Tempo (dias)", 
     ylab = "mean_value", type = "all")

C6_C150_G4 <- drm(mean_value~ tempo, data = C6_C150 , fct =  G.4())
plot(C6_C150_G4, broken = TRUE, gridsize = 100, xlab = "Tempo (dias)", 
     ylab = "mean_value", type = "all")

C7_C150_G4 <- drm(mean_value~ tempo, data = C7_C150 , fct =  G.4())
plot(C7_C150_G4, broken = TRUE, gridsize = 100, xlab = "Tempo (dias)", 
     ylab = "mean_value", type = "all")

C44_C150_G4 <- drm(mean_value~ tempo, data = C44_C150 , fct =  G.4())
plot(C44_C150_G4, broken = TRUE, gridsize = 100, xlab = "Tempo (dias)", 
     ylab = "mean_value", type = "all")

C58_C150_G4 <- drm(mean_value~ tempo, data = C58_C150 , fct =  G.4())
plot(C58_C150_G4, broken = TRUE, gridsize = 100, xlab = "Tempo (dias)", 
     ylab = "mean_value", type = "all")


C67_C150_G4 <- drm(mean_value~ tempo, data = C67_C150 , fct =  G.4())
plot(C67_C150_G4, broken = TRUE, gridsize = 100, xlab = "Tempo (dias)", 
     ylab = "mean_value", type = "all")

C73_C150_G4 <- drm(mean_value~ tempo, data = C73_C150 , fct =  G.4())
plot(C73_C150_G4, broken = TRUE, gridsize = 100, xlab = "Tempo (dias)", 
     ylab = "mean_value", type = "all")

C85_C150_G4 <- drm(mean_value~ tempo, data = C85_C150 , fct =  G.4())
plot(C85_C150_G4, broken = TRUE, gridsize = 100, xlab = "Tempo (dias)", 
     ylab = "mean_value", type = "all")

C89_C150_G4 <- drm(mean_value~ tempo, data = C89_C150 , fct =  G.4())
plot(C89_C150_G4, broken = TRUE, gridsize = 100, xlab = "Tempo (dias)", 
     ylab = "mean_value", type = "all")

newdata_REF_C150 <- expand.grid(tempo = seq(min(REF_C150$tempo), max(REF_C150$tempo), length.out = 200))
newdata_REF_C150$abs <- predict(REF_C150_G4, newdata_REF_C150, type = 'response')
newdata_REF_C150$pop <- "REF"

newdata_GSH1_C150 <- expand.grid(tempo = seq(min(GSH1_C150$tempo), max(GSH1_C150$tempo), length.out = 200))
newdata_GSH1_C150$abs <- predict(GSH1_C150_G4, newdata_GSH1_C150, type = 'response')
newdata_GSH1_C150$pop <- "GSH1"

newdata_C6_C150 <- expand.grid(tempo = seq(min(C6_C150$tempo), max(C6_C150$tempo), length.out = 200))
newdata_C6_C150$abs <- predict(C6_C150_G4, newdata_C6_C150, type = 'response')
newdata_C6_C150$pop <- "C6"

newdata_C7_C150 <- expand.grid(tempo = seq(min(C7_C150$tempo), max(C7_C150$tempo), length.out = 200))
newdata_C7_C150$abs <- predict(C7_C150_G4, newdata_C7_C150, type = 'response')
newdata_C7_C150$pop <- "C7"

newdata_C44_C150 <- expand.grid(tempo = seq(min(C44_C150$tempo), max(C44_C150$tempo), length.out = 200))
newdata_C44_C150$abs <- predict(C44_C150_G4, newdata_C44_C150, type = 'response')
newdata_C44_C150$pop <- "C44"

newdata_C58_C150 <- expand.grid(tempo = seq(min(C58_C150$tempo), max(C58_C150$tempo), length.out = 200))
newdata_C58_C150$abs <- predict(C58_C150_G4, newdata_C58_C150, type = 'response')
newdata_C58_C150$pop <- "C58"

newdata_C67_C150 <- expand.grid(tempo = seq(min(C67_C150$tempo), max(C67_C150$tempo), length.out = 200))
newdata_C67_C150$abs <- predict(C67_C150_G4, newdata_C67_C150, type = 'response')
newdata_C67_C150$pop <- "C67"

newdata_C73_C150 <- expand.grid(tempo = seq(min(C73_C150$tempo), max(C73_C150$tempo), length.out = 200))
newdata_C73_C150$abs <- predict(C73_C150_G4, newdata_C73_C150, type = 'response')
newdata_C73_C150$pop <- "C73"

newdata_C85_C150 <- expand.grid(tempo = seq(min(C85_C150$tempo), max(C85_C150$tempo), length.out = 200))
newdata_C85_C150$abs <- predict(C85_C150_G4, newdata_C85_C150, type = 'response')
newdata_C85_C150$pop <- "C85"

newdata_C89_C150 <- expand.grid(tempo = seq(min(C89_C150$tempo), max(C89_C150$tempo), length.out = 200))
newdata_C89_C150$abs <- predict(C89_C150_G4, newdata_C89_C150, type = 'response')
newdata_C89_C150$pop <- "C89"

## Exploring predicted model

GrowthCurves_C150 <- full_join(REF_C150, C6_C150)
GrowthCurves_C150 <- full_join(GrowthCurves_C150, GSH1_C150)
GrowthCurves_C150 <- full_join(GrowthCurves_C150, C7_C150)
GrowthCurves_C150 <- full_join(GrowthCurves_C150, C44_C150)
GrowthCurves_C150 <- full_join(GrowthCurves_C150, C58_C150)
GrowthCurves_C150 <- full_join(GrowthCurves_C150, C67_C150)
GrowthCurves_C150 <- full_join(GrowthCurves_C150, C73_C150)
GrowthCurves_C150 <- full_join(GrowthCurves_C150, C85_C150)
GrowthCurves_C150 <- full_join(GrowthCurves_C150, C89_C150)

newdata_full <- rbind.data.frame(newdata_REF_C150, newdata_C6_C150)
newdata_full <- rbind(newdata_full, newdata_GSH1_C150)
newdata_full <- rbind(newdata_full, newdata_C7_C150)
newdata_full <- rbind(newdata_full, newdata_C44_C150)
newdata_full <- rbind(newdata_full, newdata_C58_C150)
newdata_full <- rbind(newdata_full, newdata_C67_C150)
newdata_full <- rbind(newdata_full, newdata_C73_C150)
newdata_full <- rbind(newdata_full, newdata_C85_C150)
newdata_full <- rbind(newdata_full, newdata_C89_C150)

DoseResponseCurves_01 <- ggplot() +
  geom_point(data = REF_C150, aes(x = tempo, y = mean_value)) +
  geom_line(data = newdata_REF_C150, aes(x = tempo, y = newdata_REF_C150$abs)) +
  geom_point(data = GSH1_C150, aes(x = tempo, y = mean_value)) +
  geom_line(data = newdata_GSH1_C150, aes(x = tempo, y = newdata_GSH1_C150$abs))+
  geom_point(data = C6_C150, aes(x = tempo, y = mean_value)) +
  geom_line(data = newdata_C6_C150, aes(x = tempo, y = newdata_C6_C150$abs))+
  geom_point(data = C7_C150, aes(x = tempo, y = mean_value)) +
  geom_line(data = newdata_C7_C150, aes(x = tempo, y = newdata_C7_C150$abs)) +
  geom_point(data = C44_C150, aes(x = tempo, y = mean_value)) +
  geom_line(data = newdata_C44_C150
            , aes(x = tempo, y = newdata_C44_C150$abs))+
  geom_point(data = C58_C150, aes(x = tempo, y = mean_value)) +
  geom_line(data = newdata_C58_C150, aes(x = tempo, y = newdata_C58_C150$abs))+
  geom_point(data = C67_C150, aes(x = tempo, y = mean_value)) +
  geom_line(data = newdata_C67_C150, aes(x = tempo, y =  newdata_C67_C150$abs )) +
  geom_point(data = C73_C150, aes(x = tempo, y = mean_value)) +
  geom_line(data = newdata_C73_C150, aes(x = tempo, y =newdata_C73_C150$abs )) +
  geom_point(data = C85_C150, aes(x = tempo, y = mean_value)) +
  geom_line(data = newdata_C85_C150, aes(x = tempo, y = newdata_C85_C150$abs))+
  geom_point(data = C89_C150, aes(x = tempo, y = mean_value)) +
  geom_line(data = newdata_C89_C150, aes(x = tempo, y = newdata_C89_C150$abs))+
  ggtitle("Promastigotes Growth") +
  labs(x = "Tempo (dias)", y = "Abs") +
  theme_bw() +
  theme( plot.title = element_text(size = 14, face = "bold"),
         axis.text.x = element_text(size = 10),
         axis.text.y = element_text(size = 10),
         axis.title.x = element_text(size = 15),
         axis.title.y = element_text(size = 15)) +
  facet_wrap(~pop)

DoseResponseCurves_01
ggsave("Figures/01_DoseResponseCurves_CONC_150.png")

summary(REF_C150_G4)
# Extracting measures

COEFICIENTS <- c("GrowthRate", "c", "d", "e")
modelREF <- summary(REF_C150_G4)
confintREF <- confint(REF_C150_G4)
modelREF_coef <- as.data.frame(modelREF$coefficients)
modelREF_conf <- as.data.frame(confintREF)
modelREF_SUM <- cbind(modelREF_coef, modelREF_conf)
modelREF_SUM$coef <- COEFICIENTS 
modelREF_SUM$pop <- rep("REF_C150", 4)

COEFICIENTS <- c("GrowthRate", "c", "d", "e")
modelGSH1 <- summary(GSH1_C150_G4)
confintGSH1 <- confint(GSH1_C150_G4)
modelGSH1_coef <- as.data.frame(modelGSH1$coefficients)
modelGSH1_conf <- as.data.frame(confintGSH1)
modelGSH1_SUM <- cbind(modelGSH1_coef, modelGSH1_conf)
modelGSH1_SUM$coef <- COEFICIENTS 
modelGSH1_SUM$pop <- rep("GSH1_C150", 4)

COEFICIENTS <- c("GrowthRate", "c", "d", "e")
modelC6 <- summary(C6_C150_G4)
confintC6 <- confint(C6_C150_G4)
modelC6_coef <- as.data.frame(modelC6$coefficients)
modelC6_conf <- as.data.frame(confintC6)
modelC6_SUM <- cbind(modelC6_coef, modelC6_conf)
modelC6_SUM$coef <- COEFICIENTS 
modelC6_SUM$pop <- rep("C6_C150", 4)

COEFICIENTS <- c("GrowthRate", "c", "d", "e")
modelC7 <- summary(C7_C150_G4)
confintC7 <- confint(C7_C150_G4)
modelC7_coef <- as.data.frame(modelC7$coefficients)
modelC7_conf <- as.data.frame(confintC7)
modelC7_SUM <- cbind(modelC7_coef, modelC7_conf)
modelC7_SUM$coef <- COEFICIENTS 
modelC7_SUM$pop <- rep("C7_C150", 4)

COEFICIENTS <- c("GrowthRate", "c", "d", "e")
modelC44 <- summary(C44_C150_G4)
confintC44 <- confint(C44_C150_G4)
modelC44_coef <- as.data.frame(modelC44$coefficients)
modelC44_conf <- as.data.frame(confintC44)
modelC44_SUM <- cbind(modelC44_coef, modelC44_conf)
modelC44_SUM$coef <- COEFICIENTS 
modelC44_SUM$pop <- rep("C44_C150", 4)

COEFICIENTS <- c("GrowthRate", "c", "d", "e")
modelC58 <- summary(C58_C150_G4)
confintC58 <- confint(C58_C150_G4)
modelC58_coef <- as.data.frame(modelC58$coefficients)
modelC58_conf <- as.data.frame(confintC58)
modelC58_SUM <- cbind(modelC58_coef, modelC58_conf)
modelC58_SUM$coef <- COEFICIENTS 
modelC58_SUM$pop <- rep("C58_C150", 4)

COEFICIENTS <- c("GrowthRate", "c", "d", "e")
modelC67 <- summary(C67_C150_G4)
confintC67 <- confint(C67_C150_G4)
modelC67_coef <- as.data.frame(modelC67$coefficients)
modelC67_conf <- as.data.frame(confintC67)
modelC67_SUM <- cbind(modelC67_coef, modelC67_conf)
modelC67_SUM$coef <- COEFICIENTS 
modelC67_SUM$pop <- rep("C67_C150", 4)

COEFICIENTS <- c("GrowthRate", "c", "d", "e")
modelC73 <- summary(C73_C150_G4)
confintC73 <- confint(C73_C150_G4)
modelC73_coef <- as.data.frame(modelC73$coefficients)
modelC73_conf <- as.data.frame(confintC73)
modelC73_SUM <- cbind(modelC73_coef, modelC73_conf)
modelC73_SUM$coef <- COEFICIENTS 
modelC73_SUM$pop <- rep("C73_C150", 4)

COEFICIENTS <- c("GrowthRate", "c", "d", "e")
modelC85 <- summary(C85_C150_G4)
confintC85 <- confint(C85_C150_G4)
modelC85_coef <- as.data.frame(modelC85$coefficients)
modelC85_conf <- as.data.frame(confintC85)
modelC85_SUM <- cbind(modelC85_coef, modelC85_conf)
modelC85_SUM$coef <- COEFICIENTS 
modelC85_SUM$pop <- rep("C85_C150", 4)

COEFICIENTS <- c("GrowthRate", "c", "d", "e")
modelC89 <- summary(C89_C150_G4)
confintC89 <- confint(C89_C150_G4)
modelC89_coef <- as.data.frame(modelC89$coefficients)
modelC89_conf <- as.data.frame(confintC89)
modelC89_SUM <- cbind(modelC89_coef, modelC89_conf)
modelC89_SUM$coef <- COEFICIENTS 
modelC89_SUM$pop <- rep("C89_C150", 4)


summary_DR <- rbind(modelREF_SUM,modelGSH1_SUM,
                    modelC7_SUM, modelC6_SUM, modelC44_SUM,
                    modelC58_SUM, modelC67_SUM, modelC73_SUM,
                    modelC85_SUM, modelC89_SUM)

write.csv(summary_DR , file = "Docs/summary_GC_GSH1_C150.csv", row.names = FALSE)

# Models comparisons

m1<-drm(mean_value ~ tempo, pop, data = DataGC, 
        fct =G.4(names = c("GrowthRate", "c", "d", "e")))
m2<-drm(mean_value ~ tempo, data = DataGC, 
        fct =G.4(names = c("GrowthRate", "c", "d", "e")))

anova(m1,m2) 

DoseResponseCurves_02 <- ggplot() +
  geom_point(data = REF_C150, aes(x = tempo, y = mean_value, color = "REF")) +
  geom_line(data = newdata_REF_C150, aes(x = tempo, y = newdata_REF_C150$abs, color = "REF")) +
  geom_point(data = GSH1_C150, aes(x = tempo, y = mean_value, color = "GSH1")) +
  geom_line(data = newdata_GSH1_C150, aes(x = tempo, y = newdata_GSH1_C150$abs, color = "GSH1")) +
  geom_point(data = C6_C150, aes(x = tempo, y = mean_value, color = "C6")) +
  geom_line(data = newdata_C6_C150, aes(x = tempo, y = newdata_C6_C150$abs, color = "C6")) +
  geom_point(data = C7_C150, aes(x = tempo, y = mean_value, color = "C7")) +
  geom_line(data = newdata_C7_C150, aes(x = tempo, y = newdata_C7_C150$abs, color = "C7")) +
  geom_point(data = C44_C150, aes(x = tempo, y = mean_value, color = "C44")) +
  geom_line(data = newdata_C44_C150, aes(x = tempo, y = newdata_C44_C150$abs, color = "C44")) +
  geom_point(data = C58_C150, aes(x = tempo, y = mean_value, color = "C58")) +
  geom_line(data = newdata_C58_C150, aes(x = tempo, y = newdata_C58_C150$abs, color = "C58")) +
  geom_point(data = C67_C150, aes(x = tempo, y = mean_value, color = "C67")) +
  geom_line(data = newdata_C67_C150, aes(x = tempo, y = newdata_C67_C150$abs, color = "C67")) +
  geom_point(data = C73_C150, aes(x = tempo, y = mean_value, color = "C73")) +
  geom_line(data = newdata_C73_C150, aes(x = tempo, y = newdata_C73_C150$abs, color = "C73")) +
  geom_point(data = C85_C150, aes(x = tempo, y = mean_value, color = "C85")) +
  geom_line(data = newdata_C85_C150, aes(x = tempo, y = newdata_C85_C150$abs, color = "C85")) +
  geom_point(data = C89_C150, aes(x = tempo, y = mean_value, color = "C89")) +
  geom_line(data = newdata_C89_C150, aes(x = tempo, y = newdata_C89_C150$abs, color = "C89")) +
  ggtitle("Promastigotes Growth") +
  labs(x = "Tempo (dias)", y = "Abs", color = "População") +
  theme_bw() +
  theme(
    plot.title = element_text(size = 14, face = "bold"),
    axis.text.x = element_text(size = 14, face = "bold"),
    axis.text.y = element_text(size = 14, face = "bold"),
    axis.title.x = element_text(size = 14, face = "bold"),
    axis.title.y = element_text(size = 14, face = "bold")
  )

DoseResponseCurves_02

ggsave("Figures/02_DoseResponseCurves_CONC_150.png")

DoseResponseCurves_02+ labs(color = "Populações", size = 25)+
  theme(plot.title = element_text(size = 20, face = "bold"),
        axis.text.x = element_text(size = 20, face = "bold"),
        axis.text.y = element_text(size = 20, face = "bold"),
        axis.title.x = element_text(size = 20, face = "bold"),
        axis.title.y = element_text(size = 20, face = "bold"),
        legend.text = element_text(size = 20),
        legend.title = element_text(size = 25,  face = "bold"),
        legend.position = c(0.95,0.8),   
        legend.justification = "right",
        legend.box.background = element_blank())

ggsave("Figures/11_DoseResponseCurves.png")


#  Fit diagnostic
plot1 <- ggplot(REF_C150, aes(x = fitted(REF_C150_G4), y = mean_value)) +
  geom_point(pch = 16, cex = 0.7) +
  geom_abline(intercept = 0, slope = 1) +
  labs(title = " REF - G4 - Ajustados x Observados", x = "Valores ajustados", 
       y = "Valores observados") +
  theme_minimal()

plot2 <- ggplot(REF_C150, aes(x = fitted(REF_C150_G4), y = residuals(REF_C150_G4, type = "studentised"))) +
  geom_point(pch = 16, cex = 0.7) +
  geom_smooth(method = "loess", se = FALSE, color = "blue") +
  geom_hline(yintercept = 0, linetype = "dashed") +
  labs(title = "REF - G4 - Resíduos x Ajustados", x = "Valores ajustados", 
       y = "Resíduos") +
  theme_minimal()

plot3 <- ggplot(REF_C150, aes(sample = residuals(REF_C150_G4, type = "studentised"))) +
  geom_qq() +
  geom_qq_line() +
  labs(title = "G4", x = "Quantis teóricos", y = "Resíduos") +
  theme_minimal()

dfit_plot <- grid.arrange(plot1, plot2, plot3, ncol = 3)


ggsave("Figures/Fit_REF.png", plot)

plot1 <- ggplot(GSH1, aes(x = fitted(GSH1_C150_G4), y = abs)) +
  geom_point(pch = 16, cex = 0.7) +
  geom_abline(intercept = 0, slope = 1) +
  labs(title = " GSH1 - G4 - Ajustados x Observados", x = "Valores ajustados", 
       y = "Valores observados") +
  theme_minimal()

plot2 <- ggplot(GSH1, aes(x = fitted(GSH1_C150_G4), y = residuals(GSH1_C150_G4, type = "studentised"))) +
  geom_point(pch = 16, cex = 0.7) +
  geom_smooth(method = "loess", se = FALSE, color = "blue") +
  geom_hline(yintercept = 0, linetype = "dashed") +
  labs(title = "GSH1 - G4 - Resíduos x Ajustados", x = "Valores ajustados", 
       y = "Resíduos") +
  theme_minimal()

plot3 <- ggplot(GSH1, aes(sample = residuals(GSH1_C150_G4, type = "studentised"))) +
  geom_qq() +
  geom_qq_line() +
  labs(title = "G4", x = "Quantis teóricos", y = "Resíduos") +
  theme_minimal()

dfit_plot <- grid.arrange(plot1, plot2, plot3, ncol = 3)

ggsave("Figures/Fit_GSH1.png", plot)

plot1 <- ggplot(C6, aes(x = fitted(C6_C150_G4), y = abs)) +
  geom_point(pch = 16, cex = 0.7) +
  geom_abline(intercept = 0, slope = 1) +
  labs(title = " C6 - G4 - Ajustados x Observados", x = "Valores ajustados", 
       y = "Valores observados") +
  theme_minimal()

plot2 <- ggplot(C6, aes(x = fitted(C6_C150_G4), y = residuals(C6_C150_G4, type = "studentised"))) +
  geom_point(pch = 16, cex = 0.7) +
  geom_smooth(method = "loess", se = FALSE, color = "blue") +
  geom_hline(yintercept = 0, linetype = "dashed") +
  labs(title = "C6 - G4 - Resíduos x Ajustados", x = "Valores ajustados", 
       y = "Resíduos") +
  theme_minimal()

plot3 <- ggplot(C6, aes(sample = residuals(C6_C150_G4, type = "studentised"))) +
  geom_qq() +
  geom_qq_line() +
  labs(title = "G4", x = "Quantis teóricos", y = "Resíduos") +
  theme_minimal()

dfit_plot <- grid.arrange(plot1, plot2, plot3, ncol = 3)

ggsave("Figures/Fit_C6.png", plot)

plot1 <- ggplot(C7, aes(x = fitted(C7_C150_G4), y = abs)) +
  geom_point(pch = 16, cex = 0.7) +
  geom_abline(intercept = 0, slope = 1) +
  labs(title = " C7 - G4 - Ajustados x Observados", x = "Valores ajustados", 
       y = "Valores observados") +
  theme_minimal()

plot2 <- ggplot(C7, aes(x = fitted(C7_C150_G4), y = residuals(C7_C150_G4, type = "studentised"))) +
  geom_point(pch = 16, cex = 0.7) +
  geom_smooth(method = "loess", se = FALSE, color = "blue") +
  geom_hline(yintercept = 0, linetype = "dashed") +
  labs(title = "C7 - G4 - Resíduos x Ajustados", x = "Valores ajustados", 
       y = "Resíduos") +
  theme_minimal()

plot3 <- ggplot(C7, aes(sample = residuals(C7_C150_G4, type = "studentised"))) +
  geom_qq() +
  geom_qq_line() +
  labs(title = "G4", x = "Quantis teóricos", y = "Resíduos") +
  theme_minimal()

dfit_plot <- grid.arrange(plot1, plot2, plot3, ncol = 3)

ggsave("Figures/Fit_C7.png", plot)

plot1 <- ggplot(C44, aes(x = fitted(C44_C150_G4), y = abs)) +
  geom_point(pch = 16, cex = 0.7) +
  geom_abline(intercept = 0, slope = 1) +
  labs(title = " C44 - G4 - Ajustados x Observados", x = "Valores ajustados", 
       y = "Valores observados") +
  theme_minimal()

plot2 <- ggplot(C44, aes(x = fitted(C44_C150_G4), y = residuals(C44_C150_G4, type = "studentised"))) +
  geom_point(pch = 16, cex = 0.7) +
  geom_smooth(method = "loess", se = FALSE, color = "blue") +
  geom_hline(yintercept = 0, linetype = "dashed") +
  labs(title = "C44 - G4 - Resíduos x Ajustados", x = "Valores ajustados", 
       y = "Resíduos") +
  theme_minimal()

plot3 <- ggplot(C44, aes(sample = residuals(C44_C150_G4, type = "studentised"))) +
  geom_qq() +
  geom_qq_line() +
  labs(title = "G4", x = "Quantis teóricos", y = "Resíduos") +
  theme_minimal()

dfit_plot <- grid.arrange(plot1, plot2, plot3, ncol = 3)

ggsave("Figures/Fit_C44.png", plot)

plot1 <- ggplot(C58, aes(x = fitted(C58_C150_G4), y = abs)) +
  geom_point(pch = 16, cex = 0.7) +
  geom_abline(intercept = 0, slope = 1) +
  labs(title = " C58 - G4 - Ajustados x Observados", x = "Valores ajustados", 
       y = "Valores observados") +
  theme_minimal()

plot2 <- ggplot(C58, aes(x = fitted(C58_C150_G4), y = residuals(C58_C150_G4, type = "studentised"))) +
  geom_point(pch = 16, cex = 0.7) +
  geom_smooth(method = "loess", se = FALSE, color = "blue") +
  geom_hline(yintercept = 0, linetype = "dashed") +
  labs(title = "C58 - G4 - Resíduos x Ajustados", x = "Valores ajustados", 
       y = "Resíduos") +
  theme_minimal()

plot3 <- ggplot(C58, aes(sample = residuals(C58_C150_G4, type = "studentised"))) +
  geom_qq() +
  geom_qq_line() +
  labs(title = "G4", x = "Quantis teóricos", y = "Resíduos") +
  theme_minimal()

dfit_plot <- grid.arrange(plot1, plot2, plot3, ncol = 3)

ggsave("Figures/Fit_C58.png", plot)

plot1 <- ggplot(C67, aes(x = fitted(C67_C150_G4), y = abs)) +
  geom_point(pch = 16, cex = 0.7) +
  geom_abline(intercept = 0, slope = 1) +
  labs(title = " C67 - G4 - Ajustados x Observados", x = "Valores ajustados", 
       y = "Valores observados") +
  theme_minimal()

plot2 <- ggplot(C67, aes(x = fitted(C67_C150_G4), y = residuals(C67_C150_G4, type = "studentised"))) +
  geom_point(pch = 16, cex = 0.7) +
  geom_smooth(method = "loess", se = FALSE, color = "blue") +
  geom_hline(yintercept = 0, linetype = "dashed") +
  labs(title = "C67 - G4 - Resíduos x Ajustados", x = "Valores ajustados", 
       y = "Resíduos") +
  theme_minimal()

plot3 <- ggplot(C67, aes(sample = residuals(C67_C150_G4, type = "studentised"))) +
  geom_qq() +
  geom_qq_line() +
  labs(title = "G4", x = "Quantis teóricos", y = "Resíduos") +
  theme_minimal()

dfit_plot <- grid.arrange(plot1, plot2, plot3, ncol = 3)

ggsave("Figures/Fit_C67.png", plot)


plot1 <- ggplot(C73, aes(x = fitted(C73_C150_G4), y = abs)) +
  geom_point(pch = 16, cex = 0.7) +
  geom_abline(intercept = 0, slope = 1) +
  labs(title = " C73 - G4 - Ajustados x Observados", x = "Valores ajustados", 
       y = "Valores observados") +
  theme_minimal()

plot2 <- ggplot(C73, aes(x = fitted(C73_C150_G4), y = residuals(C73_C150_G4, type = "studentised"))) +
  geom_point(pch = 16, cex = 0.7) +
  geom_smooth(method = "loess", se = FALSE, color = "blue") +
  geom_hline(yintercept = 0, linetype = "dashed") +
  labs(title = "C73 - G4 - Resíduos x Ajustados", x = "Valores ajustados", 
       y = "Resíduos") +
  theme_minimal()

plot3 <- ggplot(C73, aes(sample = residuals(C73_C150_G4, type = "studentised"))) +
  geom_qq() +
  geom_qq_line() +
  labs(title = "G4", x = "Quantis teóricos", y = "Resíduos") +
  theme_minimal()

dfit_plot <- grid.arrange(plot1, plot2, plot3, ncol = 3)

ggsave("Figures/Fit_C73.png", plot)

plot1 <- ggplot(C85, aes(x = fitted(C85_C150_G4), y = abs)) +
  geom_point(pch = 16, cex = 0.7) +
  geom_abline(intercept = 0, slope = 1) +
  labs(title = " C85 - G4 - Ajustados x Observados", x = "Valores ajustados", 
       y = "Valores observados") +
  theme_minimal()

plot2 <- ggplot(C85, aes(x = fitted(C85_C150_G4), y = residuals(C85_C150_G4, type = "studentised"))) +
  geom_point(pch = 16, cex = 0.7) +
  geom_smooth(method = "loess", se = FALSE, color = "blue") +
  geom_hline(yintercept = 0, linetype = "dashed") +
  labs(title = "C85 - G4 - Resíduos x Ajustados", x = "Valores ajustados", 
       y = "Resíduos") +
  theme_minimal()

plot3 <- ggplot(C85, aes(sample = residuals(C85_C150_G4, type = "studentised"))) +
  geom_qq() +
  geom_qq_line() +
  labs(title = "G4", x = "Quantis teóricos", y = "Resíduos") +
  theme_minimal()

dfit_plot <- grid.arrange(plot1, plot2, plot3, ncol = 3)

ggsave("Figures/Fit_C85.png", plot)


plot1 <- ggplot(C89, aes(x = fitted(C89_C150_G4), y = abs)) +
  geom_point(pch = 16, cex = 0.7) +
  geom_abline(intercept = 0, slope = 1) +
  labs(title = " C89 - G4 - Ajustados x Observados", x = "Valores ajustados", 
       y = "Valores observados") +
  theme_minimal()

plot2 <- ggplot(C89, aes(x = fitted(C89_C150_G4), y = residuals(C89_C150_G4, type = "studentised"))) +
  geom_point(pch = 16, cex = 0.7) +
  geom_smooth(method = "loess", se = FALSE, color = "blue") +
  geom_hline(yintercept = 0, linetype = "dashed") +
  labs(title = "C89 - G4 - Resíduos x Ajustados", x = "Valores ajustados", 
       y = "Resíduos") +
  theme_minimal()

plot3 <- ggplot(C89, aes(sample = residuals(C89_C150_G4, type = "studentised"))) +
  geom_qq() +
  geom_qq_line() +
  labs(title = "G4", x = "Quantis teóricos", y = "Resíduos") +
  theme_minimal()

dfit_plot <- grid.arrange(plot1, plot2, plot3, ncol = 3)

ggsave("Figures/Fit_C89.png", plot)

plot1 <- ggplot(C76, aes(x = fitted(C76_C150_G4), y = abs)) +
  geom_point(pch = 16, cex = 0.7) +
  geom_abline(intercept = 0, slope = 1) +
  labs(title = " C76 - G4 - Ajustados x Observados", x = "Valores ajustados", 
       y = "Valores observados") +
  theme_minimal()

plot2 <- ggplot(C76, aes(x = fitted(C76_C150_G4), y = residuals(C76_C150_G4, type = "studentised"))) +
  geom_point(pch = 16, cex = 0.7) +
  geom_smooth(method = "loess", se = FALSE, color = "blue") +
  geom_hline(yintercept = 0, linetype = "dashed") +
  labs(title = "C76 - G4 - Resíduos x Ajustados", x = "Valores ajustados", 
       y = "Resíduos") +
  theme_minimal()

plot3 <- ggplot(C76, aes(sample = residuals(C76_C150_G4, type = "studentised"))) +
  geom_qq() +
  geom_qq_line() +
  labs(title = "G4", x = "Quantis teóricos", y = "Resíduos") +
  theme_minimal()

dfit_plot <- grid.arrange(plot1, plot2, plot3, ncol = 3)

ggsave("Figures/Fit_C76.png", plot)

plot1 <- ggplot(C67p, aes(x = fitted(C67p_C150_G4), y = abs)) +
  geom_point(pch = 16, cex = 0.7) +
  geom_abline(intercept = 0, slope = 1) +
  labs(title = " C67p - G4 - Ajustados x Observados", x = "Valores ajustados", 
       y = "Valores observados") +
  theme_minimal()

plot2 <- ggplot(C67p, aes(x = fitted(C67p_C150_G4), y = residuals(C67p_C150_G4, type = "studentised"))) +
  geom_point(pch = 16, cex = 0.7) +
  geom_smooth(method = "loess", se = FALSE, color = "blue") +
  geom_hline(yintercept = 0, linetype = "dashed") +
  labs(title = "C67p - G4 - Resíduos x Ajustados", x = "Valores ajustados", 
       y = "Resíduos") +
  theme_minimal()

plot3 <- ggplot(C67p, aes(sample = residuals(C67p_C150_G4, type = "studentised"))) +
  geom_qq() +
  geom_qq_line() +
  labs(title = "G4", x = "Quantis teóricos", y = "Resíduos") +
  theme_minimal()

dfit_plot <- grid.arrange(plot1, plot2, plot3, ncol = 3)

ggsave("Figures/Fit_C67p.png", plot)

plot1 <- ggplot(C68, aes(x = fitted(C68_C150_G4), y = abs)) +
  geom_point(pch = 16, cex = 0.7) +
  geom_abline(intercept = 0, slope = 1) +
  labs(title = " C68 - G4 - Ajustados x Observados", x = "Valores ajustados", 
       y = "Valores observados") +
  theme_minimal()

plot2 <- ggplot(C68, aes(x = fitted(C68_C150_G4), y = residuals(C68_C150_G4, type = "studentised"))) +
  geom_point(pch = 16, cex = 0.7) +
  geom_smooth(method = "loess", se = FALSE, color = "blue") +
  geom_hline(yintercept = 0, linetype = "dashed") +
  labs(title = "C68 - G4 - Resíduos x Ajustados", x = "Valores ajustados", 
       y = "Resíduos") +
  theme_minimal()

plot3 <- ggplot(C68, aes(sample = residuals(C68_C150_G4, type = "studentised"))) +
  geom_qq() +
  geom_qq_line() +
  labs(title = "G4", x = "Quantis teóricos", y = "Resíduos") +
  theme_minimal()

dfit_plot <- grid.arrange(plot1, plot2, plot3, ncol = 3)

ggsave("Figures/Fit_C68.png", plot)
